# src/hpc-web/src/main/resources/config/api-docs.json
{
  "swagger": "2.0",
  "info": {
    "version": "3.5.0",
...
    "title": "DME API"
  },
  "host": "hpcdmeapi.nci.nih.gov",
  "basePath": "/",
  "paths": {
...
    "/collection/download/{taskId}": {
      "get": {
        "description": "This service allows you to get the status of a collection download task. The *taskId* parameter is the task identifier returned in the *Collection Download* REST API response. \n\nThis service returns the download status of each file in DME.\n\n1. **CURL COMMAND**\n  ```\n  curl -k 'https://hpcdmeapi.nci.nih.gov:8080/collection/download/{taskId}' -D header.txt -o output.json -u <Your-NCI-User-ID>\n  ```\n\n  2. **EXAMPLE - CURL COMMAND**\n  ```\n  curl -k 'https://hpcdmeapi.nci.nih.gov:8080/collection/download/e1fc3394-f079-4f50-b3e2-8890acb0f8a5' -D header.txt -o output.json -u JohnDoe2\n  ```\n",
        "summary": "Get the status of a collection download task.",
        "tags": [
          "Get Task Status"
        ],
        "operationId": "collectionDownloadStatusByTaskId",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "taskId",
            "in": "path",
            "required": true,
            "type": "string",
            "pattern": ".*",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "**Contents of output.json File**          \n```\n{\n      \"inProgress\": false,\n      \"path\": \"/TEST_Archive/PI_Lab_JaneDoe/Project_JaneDoe\",\n      \"destinationLocation\": {\n        \"fileContainerId\": \"e99bcbb2-b6ab-11e8-8bf8-0a1d4c5c824a\",\n        \"fileId\": \"/demo/\"\n      },\n      \"destinationType\": \"GLOBUS\",\n      \"result\": \"COMPLETED\",\n      \"completedItems\": [{\n        \"dataObjectDownloadTaskId\": \"645fc12c-bcd4-4032-9f68-c00f4aeecd44\",\n        \"path\": \"/TEST_Archive/PI_Lab_JaneDoe/Project_JaneDoe/Analysis/analysis.zip\",\n        \"destinationLocation\": {\n          \"fileContainerId\": \"e99bcbb2-b6ab-11e8-8bf8-0a1d4c5c824a\",\n          \"fileId\": \"/demo//Analysis/analysis.zip\"\n        },\n        \"result\": \"COMPLETED\",\n        \"effectiveTransferSpeed\": 1,\n        \"size\": 152\n      }, {\n        \"dataObjectDownloadTaskId\": \"c9e778f6-7ac0-464a-83ff-6010602d103e\",\n        \"path\": \"/TEST_Archive/PI_Lab_JaneDoe/Project_JaneDoe/Data/data.zip\",\n        \"destinationLocation\": {\n          \"fileContainerId\": \"e99bcbb2-b6ab-11e8-8bf8-0a1d4c5c824a\",\n          \"fileId\": \"/demo//Data/data.zip\"\n        },\n        \"result\": \"COMPLETED\",\n        \"effectiveTransferSpeed\": 1,\n        \"size\": 152\n      }, {\n        \"dataObjectDownloadTaskId\": \"d42daae5-ad95-4ba7-9ef6-d6463505c422\",\n        \"path\": \"/TEST_Archive/PI_Lab_JaneDoe/Project_JaneDoe/Results/results.zip\",\n        \"destinationLocation\": {\n          \"fileContainerId\": \"e99bcbb2-b6ab-11e8-8bf8-0a1d4c5c824a\",\n          \"fileId\": \"/demo//Results/results.zip\"\n        },\n        \"result\": \"COMPLETED\",\n        \"size\": 152\n      }],\n      \"effectiveTrasnsferSpeed\": 1,\n      \"created\": \"2022-11-29T15:28:55.132-05:00\",\n      \"completed\": \"2022-11-29T15:43:30.012-05:00\"\n  }",
            "headers": {}
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Authentication Failure: \n\nHTTP/1.1 401 Unauthorized Content-Type: application/json\n\n\nJSON:\n\n{ \"errorType\": \"REQUEST_AUTHENTICATION_FAILED\",\n\n\"message\": \"Access Denied: LDAP authentication failed\",\n\n\"stackTrace\": \"…\" } "
          }
        }
      }
    },
    "/dataObject/download/{taskId}": {
      "get": {
        "description": "This service allows you to get the status of a data file download task. The *taskId* parameter is the task identifier returned in the *Data File Download* REST API response. \n\nThis service returns the download status of each data file in DME.\n1. **CURL COMMAND**\n  ```\n  curl -k 'https://hpcdmeapi.nci.nih.gov:8080/dataObject/download/{taskId}' -D header.txt -o output.json -u <Your-NCI-User-ID>\n  ```\n\n  2. **EXAMPLE - CURL COMMAND**\n  ```\n  curl -k 'https://hpcdmeapi.nci.nih.gov:8080/dataObject/download/c09b52f3-050c-4755-90c6-44d6edbd0c52' -D header.txt -o output.json -u JohnDoe2\n  ```\n",
        "summary": "Get the status of a data file download task.",
        "tags": [
          "Get Task Status"
        ],
        "operationId": "dataObjectDownloadStatusByTaskId",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "taskId",
            "in": "path",
            "required": true,
            "type": "string",
            "pattern": ".*",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "**Contents of output.json File**          \n```\n{\n  \"inProgress\": false,\n  \"path\": \"/TEST_Archive/PI_Lab_JaneDoe/Project1/data.txt\",\n  \"destinationLocation\": {\n    \"fileContainerId\": \"test01\",\n    \"fileContainerName\": \"test01\",\n    \"fileId\": \"data.txt\"\n  },\n  \"destinationType\": \"S_3\",\n  \"result\": \"COMPLETED\",\n  \"effectiveTrasnsferSpeed\": 14,\n  \"size\": 24,\n  \"created\": \"2020-07-23T16:04:52.154-04:00\",\n  \"completed\": \"2020-07-23T16:04:53.815-04:00\"\n}",
            "headers": {}
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Authentication Failure: \n\nHTTP/1.1 401 Unauthorized Content-Type: application/json\n\n\nJSON:\n\n{ \"errorType\": \"REQUEST_AUTHENTICATION_FAILED\",\n\n\"message\": \"Access Denied: LDAP authentication failed\",\n\n\"stackTrace\": \"…\" } "
          }
        }
      }
    },
...
    "/v2/collection/{path}/download": {
      "post": {
        "description": "This service allows you to to transfer a collection from DME to one of the following possible destinations:\n\n• Globus endpoint (asynchronous download)\n• AWS S3 (asynchronous download)\n• dbGaP (asynchronous download) \n\n The *path* parameter represents the logical path of the collection registered with DME. The *attributesFile.json* must have the destination information as shown in the *attributesFile.json* example below. \n\n\n\nA DME user account is a pre-requisite for downloading data from DME.  \n\nFor downloading to a Globus endpoint, if you set the optional *destinationOverwrite* indicator to ‘true’ and if the destination file exists, this service overwrites the destination file. If you omit this indicator and the destination file exists, then this service rejects the transfer request. \n\nAdditionally, for downloading to a Globus endpoint, add the **HPCDME-PROD-App-Accts-Pool-FNLCR** group to your Globus endpoint and give that group the *WRITE* permission. For details, refer to [Preparing to Use Globus](https://wiki.nci.nih.gov/x/ngLgGQ) wiki page.\n\n\n1. **CURL COMMAND**\n ```\n curl -k -H \"Content-Type: application/json\" -d @<attributesFile.json> -X POST https://hpcdmeapi.nci.nih.gov:8080/v2/collection/{path}/download -H \"Accept: application/json\" -D  <response-header> -o <response-message.json> -u <Your-NCI-User-ID>\n ```\n 2.  **attributesFile.json Example**\n \n ```\n  a. To request a download to a Globus endpoint\n        { \n          \"globusDownloadDestination\": {\n          \"destinationOverwrite\" : true,\n            \"destinationLocation\": {\n              \"fileContainerId\": \"4a3b132a-815f-11e7-8dff-2\",\n              \"fileId\": \"demo-collection-03_20_21\"\n            }\n          }\n        }\n        \n  b. To request a download to AWS S3\n        { \n          \"s3DownloadDestination\": {\n            \"destinationLocation\": {\n              \"fileContainerId\": \"demo-NCI-1\",\n              \"fileId\": \"demo-collection-03_20_21\"\n            },\n            \"account\" : {\n              \"accessKey\" : \"***\",\n              \"secretKey\" : \"***\",\n              \"region\" :\"us-east-2\"\n            }\n          }\n        }\n  \n  c. To submit collection to dbGaP \n        {\n            \"asperaDownloadDestination\": {\n                \"destinationLocation\": {\n                    \"fileContainerId\": \"protected/test\",\n                    \"fileId\": \"demo-collection_03_05_2024\"\n                },\n                \"account\": {\n                    \"host\": \"gap-submit.ncbi.nlm.nih.gov\",\n                    \"user\": \"Aspera User Id\",\n                    \"password\": \"Aspera Password\"\n                }\n            },\n            \"appendPathToDownloadDestination\": true,\n            \"appendCollectionNameToDownloadDestination\": false\n        }\n\n ```\n 3. **EXAMPLE - CURL COMMAND**\n ```\n curl -k -H \"Content-Type: application/json\" -d @attributesFile.json -X POST https://hpcdmeapi.nci.nih.gov:8080/v2/collection/TEST_Archive/TEST_Archive/Program_3_7_21/Study_3_7_21/Asset_Data_3_7_21/download -H \"Accept: application/json\" -D header.txt -o output.json -u JohnDoe2\n ```\n",
        "summary": "Download a collection.",
        "tags": [
          "Download Data"
        ],
        "operationId": "V2CollectionDownloadByPathPost",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "path",
            "in": "path",
            "required": true,
            "type": "string",
            "pattern": ".*",
            "description": ""
          },
          {
            "in": "body",
            "name": "downloadRequest",
            "description": "downloadRequest",
            "required": true,
            "schema": {
              "$ref": "#/definitions/HpcDownloadRequestDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "HTTP/1.1 200\nTBD",
            "schema": {
              "$ref": "#/definitions/HpcDownloadRequestDTO"
            },
            "headers": {}
          },
          "401": {
            "description": "Authentication Failure: \n\nHTTP/1.1 401 Unauthorized Content-Type: application/json\n\n\nJSON:\n\n{ \"errorType\": \"REQUEST_AUTHENTICATION_FAILED\",\n\n\"message\": \"Access Denied: LDAP authentication failed\",\n\n\"stackTrace\": \"…\" } "
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "HpcFileLocation": {
      "type": "object",
      "required": [
        "fileContainerId",
        "fileContainerName",
        "fileId"
      ],
      "properties": {
        "fileContainerId": {
          "type": "string"
        },
        "fileContainerName": {
          "type": "string"
        },
        "fileId": {
          "type": "string"
        }
      }
    },
...
    "HpcDownloadRequestDTO": {
      "type": "object",
      "required": [
        "destination"
      ],
      "properties": {
        "destination": {
          "$ref": "#/definitions/HpcFileLocation"
        },
        "generateDownloadRequestURL": {
          "type": "boolean"
        },
        "destinationOverwrite": {
          "type": "boolean"
        }
      },
      "xml": {
        "name": "HpcDownloadRequestDTO"
      }
    },
...
  }
}